use CSIM cell size as too close factor
