+++ src/main/java/org/streeto/scorers/TooCloseToAFutureControlScorer.java
@@ -89,7 +89,7 @@ public class TooCloseToAFutureControlScorer extends AbstractLegScorer {
     }
 
     private Double scoreRoute(double bestDistance, List<GHPoint3D> remainingControls, ResponsePath legRoute) {
-        var likelihood = legRoute.getDistance() < preferences.getMinControlSeparation() ? 1.0 : bestDistance / legRoute.getDistance();
+        var likelihood = legRoute.getDistance() < preferences.getCSIMCellSize() ? 1.0 : bestDistance / legRoute.getDistance();
         var points = legRoute.getPoints();
         var size = points.size();
         var minDist = Double.MAX_VALUE;
@@ -101,7 +101,7 @@ public class TooCloseToAFutureControlScorer extends AbstractLegScorer {
                 }
             }
         }
-        if (minDist < preferences.getMinControlSeparation()) {
+        if (minDist < preferences.getCSIMCellSize()) {
             return 1.0 - likelihood;
         } else {
             return 1.0;
